float -> double

+++ src/object.cpp
@@ -4,17 +4,17 @@
 #include <glm/gtc/constants.hpp>
 #include <cmath>
 
-static glm::vec3 sphericalToCartesian(float r, float theta, float phi) {
-    float x = r * sin(theta) * cos(phi);
-    float y = r * cos(theta);
-    float z = r * sin(theta) * sin(phi);
-    return glm::vec3(x, y, z);
+static glm::dvec3 sphericalToCartesian(double r, double theta, double phi) {
+    double x = r * sin(theta) * cos(phi);
+    double y = r * cos(theta);
+    double z = r * sin(theta) * sin(phi);
+    return glm::dvec3(x, y, z);
 }
 
-Object::Object(glm::vec3 initPosition, glm::vec3 initVelocity,
-               std::optional<float> massOpt,
-               std::optional<float> densityOpt,
-               std::optional<float> radiusOpt)
+Object::Object(glm::dvec3 initPosition, glm::dvec3 initVelocity,
+               std::optional<double> massOpt,
+               std::optional<double> densityOpt,
+               std::optional<double> radiusOpt)
 {
     position = initPosition;
     velocity = initVelocity;
@@ -32,16 +32,16 @@ Object::Object(glm::vec3 initPosition, glm::vec3 initVelocity,
     } else if (massOpt.has_value() && densityOpt.has_value()) {
         mass    = *massOpt;
         density = *densityOpt;
-        radius  = std::pow((3 * mass / density) / (4 * 3.14159265359f), 1.0f/3.0f) / 100000.0f;
+        radius  = std::pow((3.0 * mass / density) / (4.0 * 3.14159265359), 1.0/3.0) / 100000.0;
     } else if (massOpt.has_value() && radiusOpt.has_value()) {
-        float volume = (4.0f / 3.0f) * 3.14159265359f * std::pow(*radiusOpt, 3);
-        float dens   = *massOpt / volume;
+        double volume = (4.0 / 3.0) * 3.14159265359 * std::pow(*radiusOpt, 3);
+        double dens   = *massOpt / volume;
         density = dens;
         mass    = *massOpt;
         radius  = *radiusOpt;
     } else if (radiusOpt.has_value() && densityOpt.has_value()) {
-        float volume = (4.0f / 3.0f) * 3.14159265359f * std::pow(*radiusOpt, 3);
-        float m = *densityOpt * volume;
+        double volume = (4.0 / 3.0) * 3.14159265359 * std::pow(*radiusOpt, 3);
+        double m = *densityOpt * volume;
         mass    = m;
         density = *densityOpt;
         radius  = *radiusOpt;
@@ -72,27 +72,27 @@ std::vector<float> Object::Draw() {
             float phi1 = (j / (float)sectors) * 2 * glm::pi<float>();
             float phi2 = ((j+1) / (float)sectors) * 2 * glm::pi<float>();
 
-            glm::vec3 v1 = sphericalToCartesian(radius, theta1, phi1);
-            glm::vec3 v2 = sphericalToCartesian(radius, theta1, phi2);
-            glm::vec3 v3 = sphericalToCartesian(radius, theta2, phi1);
-            glm::vec3 v4 = sphericalToCartesian(radius, theta2, phi2);
+            glm::dvec3 v1 = sphericalToCartesian(radius, theta1, phi1);
+            glm::dvec3 v2 = sphericalToCartesian(radius, theta1, phi2);
+            glm::dvec3 v3 = sphericalToCartesian(radius, theta2, phi1);
+            glm::dvec3 v4 = sphericalToCartesian(radius, theta2, phi2);
 
-            // Первый треугольник 
-            vertices.insert(vertices.end(), {v1.x, v1.y, v1.z});
-            vertices.insert(vertices.end(), {v2.x, v2.y, v2.z});
-            vertices.insert(vertices.end(), {v3.x, v3.y, v3.z});
-            // Второй треугольник 
-            vertices.insert(vertices.end(), {v2.x, v2.y, v2.z});
-            vertices.insert(vertices.end(), {v4.x, v4.y, v4.z});
-            vertices.insert(vertices.end(), {v3.x, v3.y, v3.z});
+            // Первый треугольник
+            vertices.insert(vertices.end(), {static_cast<float>(v1.x), static_cast<float>(v1.y), static_cast<float>(v1.z)});
+            vertices.insert(vertices.end(), {static_cast<float>(v2.x), static_cast<float>(v2.y), static_cast<float>(v2.z)});
+            vertices.insert(vertices.end(), {static_cast<float>(v3.x), static_cast<float>(v3.y), static_cast<float>(v3.z)});
+            // Второй треугольник
+            vertices.insert(vertices.end(), {static_cast<float>(v2.x), static_cast<float>(v2.y), static_cast<float>(v2.z)});
+            vertices.insert(vertices.end(), {static_cast<float>(v4.x), static_cast<float>(v4.y), static_cast<float>(v4.z)});
+            vertices.insert(vertices.end(), {static_cast<float>(v3.x), static_cast<float>(v3.y), static_cast<float>(v3.z)});
         }
     }
     return vertices;
 }
 
-void Object::UpdatePos(float deltaTime) {
+void Object::UpdatePos(double deltaTime) {
     position += velocity * deltaTime;
-    radius = std::pow((3 * mass / density) / (4 * 3.14159265359f), 1.0f/3.0f) / 100000.0f;
+    radius = std::pow((3.0 * mass / density) / (4.0 * 3.14159265359), 1.0/3.0) / 100000.0;
 }
 
 void Object::UpdateVertices() {
@@ -101,11 +101,11 @@ void Object::UpdateVertices() {
     glBufferData(GL_ARRAY_BUFFER, vertices.size() * sizeof(float), vertices.data(), GL_STATIC_DRAW);
 }
 
-glm::vec3 Object::GetPos() const {
+glm::dvec3 Object::GetPos() const {
     return position;
 }
 
-void Object::accelerate(float x, float y, float z, float deltaTime) {
+void Object::accelerate(double x, double y, double z, double deltaTime) {
     velocity.x += x * deltaTime;
     velocity.y += y * deltaTime;
     velocity.z += z * deltaTime;
